+++ plugins/trivy/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.28.7
+* update trivy to 0.48.1
+
+## 0.28.6
+* Bump kubectl to 1.29.0
+
+## 0.28.5
+* Bump alpine to 3.19
+
 ## 0.28.4
 * Update dependencies